Reorganized folder structure.
This brings the client in line with our other web applications, and should make debugging and development easier. This is less optimal; a minified build is much faster to load, but this allows us to debug the code directly, instead of relying on source maps (and source map support is iffy at best).

This is the first step towards removing minification.